+++ js/utils.js
@@ -1,0 +1,16 @@
+// Генерирование случайного числа в диапазоне
+
+const getPositiveInteger = (min, max) => {
+  min = Math.ceil(min);
+  max = Math.floor(max);
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+
+};
+
+export {getPositiveInteger};
+
+// Вычисление длины комментария
+
+const checkStringLength = (string, stringLengthMax) => {
+  return string.length <= stringLengthMax;
+};
